Completed Part 7
This is now basically an actual game, albeit a limited one. I've changed
the system font to something that felt more appropriate (arial looks
hideous in monospace) and now there's a message log in the game window
instead of the console.

Fun stuff! Watch for the creepy log in message.