Switched to ASCII row font
After some amount of monkeying about this morning searching for an
alternate font to the IBM design, I've decided to go with it anyway. But
since I'm using the IBM font, I figured why not make full use of the
character set, so I switched to the by-rows layout so that now, the
character is the smiley!