Finished Part 12
This is mostly behind the scenes changes, revising the system for coming
up with random choices, though there's also been some rebalance changes
to the constants and initial stats for pretty much everything.

I don't actually like any of the latter changes, but most of those
numbers are going to be rewritten once the prototype is done anyway.
Right now my interest remains in completing the prototype.