The game now has an ending.
After level 7, the game now plays an ending text sequence and returns to
the main menu.

As well, there is now an established behavior for ending games.
'savegame' is deleted on a death or a win if present, so that the state
of the 'new game' menu option will now reflect whether there's an
active, playable save. This also prevents reloading a dead or won save.

I've also included a new function for handling these text 'cinematic'
sequences, in order to avoid replicating code. show_text_log takes two
arguments, text (a list of strings to play out), and img, the pic to be
drawn as the background. If img is None, it will simply draw on a black
screen.