H&H style combat now 50% complete
After a considerable amount of rewriting, there is now a more H&H-like
combat system. Player and monsters have to-hit and AC values that are
checked as normal.

I haven't yet figured out an elegant way to handle weapon/spell damages
however, so for now, attack damage is a  1d6 roll plus a bonus, and the
'spells' are still flat values.