Random damage now complete.
Thanks to the * unpacking operator, implementing random damage turned
out to be a snap!

Monsters, players, and even weapons, all can now roll random damages by
storing a list of rolldice args and then passing it to rolldice with a
leading *.

This even works for the spell constants, so they are now randomized as
well!

There is a thiiiin possibility that it's not taking the highest= option,
but initial testing doesn't seem to confirm it. May have to copy out the
code and run some tests in the interpreter.

With this implementation though, the basic engine is now complete for
milestone 0.1. Next step: setting up a binary release!