v.0.2 ready for release, now with High scores!
This update at last brings a working high score system to the game.

A local shelve file is created that takes a list of lists of the
relevant data, which can then be fed back into a routine that prepares
the data and sends it to show_text_log.

Because I wanted the extra tidbit, I've also added a new optional
parameter to Object, killed_by, which is set by a new second parameter
to Fighter.take_damage(). This can be any string, and in the case of
direct attacks is fed with monster.name so that we get 'killed by
lobsterman' or whatever in the score entry.

With that, milestone 0.2 is officially to bed. Enjoy the new binary
build, there's a lot to see!