New monster generator
The monsters now generate using a custom function
(get_monster_from_hitdice) that builds the monster object and it's stats
using the hitdice formulas in the H&H corebook.

This requires an additional dictionary with matching keys to the ones in
monster_chances, the contents of which is lists containing lists for
each of the keys listing their name, hitdice tuple, and color, in that
order.

Requiring two seperate dictionaries isn't the most ideal, but it avoids
having to rewrite the logic of random_choice and random_choice_index.

In the future, the monster_chances and monster_tables dicts can probably
be loaded from a file, to cut down on code lines.

Also, I've replaced the centipod with the blind troll in the monster
list because I felt it provided a more natural progression while we
still lack special abilities for monsters.