New monsters!
Further revised the monster table setup. It now calls a function from a
new module for these things, hhtable.py, to populate the dict (this is
basically a space saving measure).

The upshot of this is, I've now included the entire encounter list for
an underground environment as according to the H&H corebook. That means
there's now a full 20 monster types to encounter! Have fun!