Adjusted item chances
Went back to random_choice for item chances so that I can fine tune the
values better. While H&H is pretty much flat everywhere, for the chaos,
missing stuff in the handhRL lists throws off the chances tables. This
should allow me to adjust the odds better, in particular so it's not so
stingy with items.